requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Deprecated==1.2.10
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyGithub==1.53
PyJWT==1.7.1
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
